Add files via upload
IT'S ALIVEEEEE!!! I normalized the image dataset to make the values range from 0 to 1, which greatly improved the model's ability to learn the images. I also partially rebuilt the model, which also improved runtime and model complexity. After running and using the model to predict particle masses, I made some scatterplots of the true - prediction, ruled out outliers, then calculated the standard deviations for the true particle masses above 3.25 keV and below 3.25 keV. 3.25 keV is the cited point where models tend to begin to lose their predictive accuracy, so I wanted to see if that trend applied to our 2-D CNN as well. The standard deviation for below 3.25 keV was ~0.6, while the standard deviation for above 3.25 keV was ~5.3. This clearly shows that the 2-D CNN does much better on particle masses below 3.25 keV!!!!